+++ lib/src/exceptions/exceptions.dart
@@ -137,6 +137,10 @@ abstract class SoLoudCppException extends SoLoudException {
         return const SoLoudXiphLibsNotFoundCppException();
       case PlayerErrors.busIdNotFound:
         return const SoLoudBusIdNotFoundCppException();
+      case PlayerErrors.hashIsNotAPullBufferStream:
+        return const SoLoudHashIsNotAPullBufferStreamCppException();
+      case PlayerErrors.invalidPullBufferState:
+        return const SoLoudInvalidPullBufferStateCppException();
     }
   }